+++ docs/changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v1.1.0
+
+* Add [HTTP-date](https://httpwg.org/specs/rfc9110.html#http.date) value parsing for [`retryhttp.wait_from_header`][]
+* [`is_rate_limited`][`retryhttp._utils.is_rate_limited`] now determines that a request was rate limited by the presence of a `Retry-After` header. Prior to v1.1.0, this was based on the status code `429 Too Many Requests`.
+
 ## v1.0.1
 
 * Fix documentation errors.